Migrate archiving server-side code to plugins
